% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemrdata.R
\name{accountCount}
\alias{accountCount}
\title{Number of Accounts on Steem Blockchain}
\usage{
accountCount(node)
}
\arguments{
\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
number of accounts
}
\description{
Get the count of the number of accounts on the Steem Blockchain
}
\examples{
\dontrun{
accountCount()
}
}
