% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{adailyf}
\alias{adailyf}
\title{Calculate the the daily frequency.
adailyf means 'analysis of the daily frequency'.}
\usage{
adailyf(mydata, datecol, if_plot = FALSE, col = "steelblue3", ylab = "Daily posts")
}
\arguments{
\item{mydata}{A data frame with a date column.}

\item{datecol}{A chracter string of the date column name. The date column could either be Date or Character (in '\%Y-\%m-\%d' format).}

\item{if_plot}{A logic value of whether plot the time series}

\item{col}{A color for plotting}

\item{ylab}{A character string of the y label}
}
\value{
A dataframe of the daily frequency
}
\description{
Calculate the the daily frequency.
adailyf means 'analysis of the daily frequency'.
}
