% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{avotenot}
\alias{avotenot}
\title{Find which followers have not voted a post yet. avotenot means 'analysis who has not voted yet'.}
\usage{
avotenot(postlink = NA)
}
\arguments{
\item{postlink}{A character string of the link to a target post.}
}
\value{
A character vector of the name list of the followers who have not voted the target post yet.
}
\description{
Find which followers have not voted a post yet. avotenot means 'analysis who has not voted yet'.
}
\examples{
avotenot("cn/@dapeng/steemit-markdown")
}
