% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{bmd}
\alias{bmd}
\title{build markdown files from the posts}
\usage{
bmd(post_df, dest_path = "blog", post_df_source = c("appbase_api", "steemsql.com"), 
    if_yaml = FALSE, original_link = c("steemit.com", "cnsteem.com", "busy.org", 
        "steemdb.com", "steemd.com"))
}
\arguments{
\item{post_df}{A data frame of the posts.}

\item{dest_path}{A character string of the destination path for the markdown files.}

\item{post_df_source}{A character string of the data frame source.}

\item{if_yaml}{A logical value of whether the markdown files contain yaml headers.}

\item{orginal_link}{A character string vector of the orginal links to the posts.}
}
\value{
markdown files.
}
\description{
build markdown files from the posts
}
