% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{charsplit}
\alias{charsplit}
\title{Split the characters in a data frame}
\usage{
charsplit(dataframe)
}
\arguments{
\item{dataframe}{A column of a dataframe with characters to split}
}
\value{
A character vector
}
\description{
Split the characters in a data frame
}
