% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{clear_title}
\alias{clear_title}
\title{Clear the characters in the titles}
\usage{
clear_title(title)
}
\arguments{
\item{title}{The post title}
}
\value{
A clear title
}
\description{
Clear the characters in the titles
}
