% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{clearferfing}
\alias{clearferfing}
\title{Clear the format of the 'follow' information from steemdb.com.}
\usage{
clearferfing(x)
}
\arguments{
\item{x}{A character string of the follower information from steemdb.com.}
}
\value{
A character string of the formatted follower information
}
\description{
Clear the format of the 'follow' information from steemdb.com.
}
