% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{follower}
\alias{follower}
\title{A id list of an id's followers from steemdb.com}
\usage{
follower(id = "dapeng")
}
\arguments{
\item{id}{character without '@'}
}
\value{
character a dataframe of an ID's followers' info
}
\description{
A id list of an id's followers from steemdb.com
}
\examples{
follower()
}
