% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{follower_df}
\alias{follower_df}
\title{A id list of an id's followers from steemdb.com}
\usage{
follower_df(id = "dapeng")
}
\arguments{
\item{id}{character without '@'}
}
\value{
character a dataframe of an ID's followers' info
}
\description{
A id list of an id's followers from steemdb.com
}
\examples{
follower_df()
}
