% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{following}
\alias{following}
\title{The id list of an id's following  from steemdb.com}
\usage{
following(id = NA)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}
}
\value{
A dataframe of an ID's following info
}
\description{
The id list of an id's following  from steemdb.com
}
\examples{
following("dapeng")
}
