% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{followlist}
\alias{followlist}
\title{An ID's followers and following}
\usage{
followlist(id = "dapeng", method = c("steemdata.com", "steemdb.com"))
}
\arguments{
\item{id}{character without '@'}
}
\value{
a name list of an ID's followers and following
}
\description{
An ID's followers and following
}
\examples{
followlist()
}
