% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gblog}
\alias{gblog}
\title{Rename the getBlog() function from the steemRdata package}
\usage{
gblog(id = NA)
}
\arguments{
\item{id}{A Steem ID}
}
\value{
A dataframe with the ID's posts
}
\description{
Rename the getBlog() function from the steemRdata package
}
\examples{
gblog()
}
