% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gcner}
\alias{gcner}
\title{Get the CNer name list}
\usage{
gcner(date = Sys.Date())
}
\arguments{
\item{date}{A Date or character string in '\%Y-%m-%d' format}
}
\value{
a data frame of the CNer
}
\description{
Get the CNer name list
}
\examples{
gcner()
}
