% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gdelegation}
\alias{gdelegation}
\title{Get the delegation information of a Steem ID from SteemSQL
gdelegation means 'get delegation information'.}
\usage{
gdelegation(id, sql_con, if_plot = FALSE)
}
\arguments{
\item{id}{A character string of a Steem ID}

\item{sql_con}{A SQL connection}

\item{if_plot}{A logic value of whether plot the time series}
}
\value{
A data frame of the delegation information with a figure
}
\description{
Get the delegation information of a Steem ID from SteemSQL
gdelegation means 'get delegation information'.
}
\examples{
gdelegation()
}
