% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemrdata.R
\name{getSteemProperties}
\alias{getSteemProperties}
\title{Steem Proporties}
\usage{
getSteemProperties(node)
}
\arguments{
\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
steem properties
}
\description{
Get details about the latest state of the Steem Blockchain.
}
\examples{
\dontrun{
getSteemProperties()
}
}
