% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemrdata.R
\name{getTransactions}
\alias{getTransactions}
\title{Transactions made on an account}
\usage{
getTransactions(user, n, node)
}
\arguments{
\item{user}{Account To Query}

\item{n}{Number of Transactions since beginnning of account history}

\item{node}{node}
}
\value{
The First n of transactions on an account
}
\description{
This function may take some time to process depending on the size of the transaction history on an account
}
\examples{
\dontrun{
getTransactions("eroche", 100)
}
}
