% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gfollow}
\alias{gfollow}
\title{Get a name list of an ID's followers and following. gfollow means 'get an ID's follower information'.}
\usage{
gfollow(id = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), sql_con)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{sql_con}{A connection to the SteemSQL server.}
}
\value{
A name list of an ID's followers and following
}
\description{
Get a name list of an ID's followers and following. gfollow means 'get an ID's follower information'.
}
\examples{
gfollow("dapeng")
}
