% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gfollower}
\alias{gfollower}
\title{A list of an id's followers from steemdb.com. gfollower means 'get the follower information'.}
\usage{
gfollower(id = NA)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}
}
\value{
character a dataframe of an ID's followers' info
}
\description{
A list of an id's followers from steemdb.com. gfollower means 'get the follower information'.
}
\examples{
gfollower("dapeng")
}
