% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gfollowing}
\alias{gfollowing}
\title{The id list of an id's following  from steemdb.com. gfollowing means 'get the following information'.}
\usage{
gfollowing(id = NA)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}
}
\value{
A dataframe of an ID's following info
}
\description{
The id list of an id's following  from steemdb.com. gfollowing means 'get the following information'.
}
\examples{
gfollowing("dapeng")
}
