% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gid}
\alias{gid}
\title{Get an ID's detailed info. gid means 'get an ID's information'.}
\usage{
gid(id = NA, method = c("appbase_api", "steemdb.com", "steemsql.com", "steemdata.com"), 
    sql_con)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{sql_con}{A connection to the SteemSQL server.}
}
\value{
A list or a dataframe of an ID's detailed info
}
\description{
Get an ID's detailed info. gid means 'get an ID's information'.
}
\examples{
gid(id = "dapeng")
}
