% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gidpostl}
\alias{gidpostl}
\title{Obtain an ID's post hyperlinks. gidpostl means 'get an ID's post links'.}
\usage{
gidpostl(id = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), sql_con, 
    post_number = 3, site = "steemit.com")
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{post_number}{A numeric value or NA. The number of the latest posts to be obtained. If NA, the 100 latest posts will be processed.}

\item{site}{A character string of the site of the steem web UI}

\item{steemsql_connection}{A connection to the SteemSQL server.}
}
\value{
A character string vector of an ID's post hyperlinks.
}
\description{
Obtain an ID's post hyperlinks. gidpostl means 'get an ID's post links'.
}
\examples{
gidpostl()
}
