% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gidposts}
\alias{gidposts}
\title{Obtain an ID's post detailed info from steemdb.com. 'gidposts' means 'get an ID's posts'.}
\usage{
gidposts(id = NA, method = c("steemdb.com", "steemsql.com", "appbase_api", 
    "steemdata.com"), sql_con, post_number = NA)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{post_number}{A numeric value or NA. The number of the latest posts to be obtained. If NA, all the posts will be processed.}

\item{steemsql_connection}{A connection to the SteemSQL server.}
}
\value{
A data frame of an ID's post detailed info.
}
\description{
Obtain an ID's post detailed info from steemdb.com. 'gidposts' means 'get an ID's posts'.
}
\examples{
gidposts("dapeng")
}
