% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gpost}
\alias{gpost}
\title{Get the complete info of a single given post. gpost means 'get a post'.}
\usage{
gpost(postlink = NA, method = c("steemdb.com", "steemsql.com"), sql_con, 
    selected = FALSE, newline = FALSE, oldcolname)
}
\arguments{
\item{postlink}{A character of the link to a post}

\item{method}{A character string of the Steem data server to connect.}

\item{sql_con}{A connection to the SteemSQL server.}

\item{selected}{A logic value of whether return only selected info}

\item{newline}{A logic value of whether rbind the returned dataframe with an existing one}

\item{oldcolname}{A character string. If newline == TRUE, the returned dataframe is ordered accoring to oldcolname}
}
\value{
a dataframe
}
\description{
Get the complete info of a single given post. gpost means 'get a post'.
}
\examples{
gpost("cn/@dapeng/steemit-markdown")
}
