% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gposts}
\alias{gposts}
\title{Get the detailed information of given posts. gposts means 'get posts'.}
\usage{
gposts(postlinks = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), 
    sql_con)
}
\arguments{
\item{postlinks}{A character string of hyperlinks to target posts}

\item{method}{A character string of the Steem data server to connect.}

\item{sql_con}{A connection to the SteemSQL server.}
}
\value{
a dataframe of the detailed information of the given posts
}
\description{
Get the detailed information of given posts. gposts means 'get posts'.
}
\examples{
gposts(c("cn/@dapeng/xuer-sale", "utopian-io/@dapeng/steemg-four-more"))
}
