% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gur}
\alias{gur}
\title{Get the utopian review data}
\usage{
gur(id = NA, show_all = TRUE)
}
\arguments{
\item{id}{character. The ID to search in the utopian review plan.}

\item{show_all}{logic.}
}
\value{
A dataframe
}
\description{
Get the utopian review data
}
\examples{
gur()
}
