% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{hourrose}
\alias{hourrose}
\title{Hour rose plot}
\usage{
hourrose(mydata, ws = "ws", wd = "wd", ws2 = NA, wd2 = NA, ws.int = 2, angle = 30, 
    type = "default", cols = "default", grid.line = NULL, width = 1, seg = 0.9, 
    auto.text = TRUE, breaks = 4, offset = 10, paddle = FALSE, key.header = NULL, 
    key.footer = "(m/s)", key.position = "right", key = TRUE, dig.lab = 5, 
    statistic = "prop.count", pollutant = NULL, cust_labels = c(0, 6, 12, 18), 
    annotate = TRUE, border = NA, quantile_line = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
a figure
}
\description{
Hour rose plot
}
\examples{
hourrose(mydata, breaks = seq(0, 24, 1), angle = 15, key.footer = "votes")
}
