% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{idinfo}
\alias{idinfo}
\title{An ID's detailed info}
\usage{
idinfo(id = "dapeng", method = c("steemdata.com", "steemdb.com"))
}
\arguments{
\item{id}{character. Steem ID without '@'}

\item{method}{character. Steem data server.}
}
\value{
a dataframe of an ID's detailed info
}
\description{
An ID's detailed info
}
\examples{
followlist()
}
