% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{info}
\alias{info}
\title{Download data from steemdata.com}
\usage{
info(method = c("A", "P", "V"), A = "dapeng", P = "", V = "abit")
}
\arguments{
\item{method}{character. specifying what report.
\itemize{
\item A: Account Report;
\item P: Post Report;
\item V: Voter Report
}}

\item{A}{character. The name of the account for account report}

\item{P}{character. The name of the account for post report}

\item{V}{character. The name of the account for voter report}
}
\value{
a list
}
\description{
Download data from steemdata.com
}
\examples{
info()
}
