% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mirror}
\alias{mirror}
\title{Title build a blogdown site for a steem author}
\usage{
mirror(author = "dapeng", dloadall = FALSE, initial = FALSE, template = "xmin")
}
\arguments{
\item{author}{author name without @}

\item{dloadall}{if download all posts, or only download a given period}

\item{initial}{if initialize a site}

\item{template}{the hugo template}
}
\value{
a blogdown site
}
\description{
Title build a blogdown site for a steem author
}
