% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{pcner}
\alias{pcner}
\title{Plot function for the Shiny app scner}
\usage{
pcner(i, sliderplot, whwechat, zh)
}
\arguments{
\item{i}{A numeric indicator}

\item{sliderplot}{The ID of the sliderplot}

\item{whwechat}{The data from of the cner}

\item{zh}{The Chinese dictionary}
}
\value{
A diagram
}
\description{
Plot function for the Shiny app scner
}
