% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{pdate}
\alias{pdate}
\title{Plot a time series with x as Date.}
\usage{
pdate(x, y, myylab = "", mylegend = "", mycol = "darkgreen", myxlim = NULL, 
    myylim = NULL)
}
\arguments{
\item{x}{A Date vector.}

\item{y}{A numeric vector.}

\item{myylab}{A character string of the y label.}

\item{mylegend}{A character string of the legend.}

\item{mycol}{A color.}

\item{myxlim}{A Date vector of the date range of x.}
}
\value{
A time series diagram.
}
\description{
Plot a time series with x as Date.
}
