% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{phist}
\alias{phist}
\title{Plot a histogram of a vector.}
\usage{
phist(data = rnorm(1000), mybreaks = "Sturges", myxlim = NULL, myylim = NULL, 
    eightlines = TRUE, eightdigit = 0, eightcex = 0.8, eightcolors = c("red", 
        "darkgreen", "blue", "black", "purple", "gold")[c(1, 2, 3, 2, 1, 6, 6, 
        5, 4, 5)], mylegend = "", myxlab = "", return_df = FALSE, myfreq = FALSE, 
    show_n = TRUE, show_skewness = TRUE, show_density = FALSE, show_normline = FALSE, 
    show_mean = FALSE)
}
\arguments{
\item{data}{A numeric vector to plot.}

\item{mybreaks}{Breaks of the plotted bars.}

\item{myxlim}{x limit.}

\item{myylim}{y limit.}

\item{eightlines}{Whether to draw the eight lines.}

\item{eightdigit}{The digit number of the figures displayed on the diagram.}

\item{eightcex}{The character size of the figures.}

\item{eightcolors}{The colors of the eight lines.}

\item{mylegend}{The legend.}

\item{myxlab}{The x label.}

\item{return_df}{Whether return the data frame of summary.}

\item{myfreq}{Whether display the frequency,}

\item{show_n}{Whether show the sample number.}

\item{show_skewness}{Whether show the skewness test.}

\item{show_density}{Whether show the density.}

\item{show_normline}{Whether show the line of the normal distribution.}

\item{show_mean}{Whether show the mean value.}
}
\value{
A histogram diagram.
}
\description{
Plot a histogram of a vector.
}
