% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{post_df}
\alias{post_df}
\title{Get the detailed information of given posts}
\usage{
post_df(postlinks = c("cn/@dapeng/xuer-sale", "utopian-io/@dapeng/steemg-four-more"), 
    method = c("steemdb.com", "steemsql.com", "steemdata.com"), sql_id, sql_password)
}
\arguments{
\item{postlinks}{A character string of hyperlinks to target posts}

\item{method}{A character string of the Steem data server to connect.}

\item{sql_id}{A character string of the steemsql id. Only valid when \code{method == 'steemsql'}}

\item{sql_password}{A character string of the steemsql passwordid. Only valid when \code{method == 'steemsql'}}
}
\value{
a dataframe of the detailed information of the given posts
}
\description{
Get the detailed information of given posts
}
\examples{
post_df()
}
