% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{post_info}
\alias{post_info}
\title{Get the complete info of a single given post on steemdb.com}
\usage{
post_info(postlink = "cn/@dapeng/steemit-markdown", selected = FALSE, newline = FALSE, 
    oldcolname)
}
\arguments{
\item{postlink}{character. the complete hyperlink of a post on steemdb.com}

\item{selected}{logic. Whether return only selected info}

\item{newline}{logic. Whether rbind the returned dataframe with an existing one}

\item{oldcolname}{character. If newline == TRUE, the returned dataframe is ordered accoring to oldcolname}
}
\value{
a dataframe
}
\description{
Get the complete info of a single given post on steemdb.com
}
\examples{
post_info()
}
