% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{post_rmd}
\alias{post_rmd}
\title{Create a markdown file accoring to given post links from steemdb.com}
\usage{
post_rmd(rmdnr = "01", rmdtopic = "", postlinks)
}
\arguments{
\item{rmdnr}{character. a number appearing in the markdown file name}

\item{rmdtopic}{character. a topic appreaing in the markdown file name}

\item{postlinks}{character. hyperlinks of target posts}
}
\value{
a markdown file
}
\description{
Create a markdown file accoring to given post links from steemdb.com
}
\examples{
post_rmd()
}
