% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{postlink_id}
\alias{postlink_id}
\title{Obtain an ID's post hyperlinks from steemdb.com}
\usage{
postlink_id(id = "dapeng", post_number = 3)
}
\arguments{
\item{id}{character without '@'}

\item{post_number}{numeric or NA. The number of the latest posts to be obtained. If NA, all the posts will be processed.}
}
\value{
character. an ID's post hyperlinks.
}
\description{
Obtain an ID's post hyperlinks from steemdb.com
}
\examples{
post_id()
}
