% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{repcalc}
\alias{repcalc}
\title{Calculate the reputation of an ID}
\usage{
repcalc(rep)
}
\arguments{
\item{rep}{A numeric value of the raw reputation.}
}
\value{
A numeric value of the real reputation
}
\description{
Calculate the reputation of an ID
}
