% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{scner}
\alias{scner}
\title{A shiny app to display and anaylize the CNers.
scner means shiny app for CNers.}
\usage{
scner()
}
\value{
a shinyapp which can be displayed in a web browser.
}
\description{
A shiny app to display and anaylize the CNers.
scner means shiny app for CNers.
}
