% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{sfollow}
\alias{sfollow}
\title{A shiny app to display and anaylize the followers a given ID.
sfollow means shiny app for followers.}
\usage{
sfollow()
}
\value{
a shinyapp which can be displayed in a web browser.
}
\description{
A shiny app to display and anaylize the followers a given ID.
sfollow means shiny app for followers.
}
