% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{sfollow_server}
\alias{sfollow_server}
\title{Server for the Shiny app sfollow display and analysis}
\usage{
sfollow_server(input, output, session)
}
\arguments{
\item{input}{The input of the server.}

\item{output}{The output of the server.}

\item{session}{The session of the server.}
}
\value{
A server function
}
\description{
Server for the Shiny app sfollow display and analysis
}
