% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{sposts}
\alias{sposts}
\title{A shiny app to display and anaylize the posts of a given ID.
sposts means shiny app for posts.}
\usage{
sposts()
}
\value{
a shinyapp which can be displayed in a web browser.
}
\description{
A shiny app to display and anaylize the posts of a given ID.
sposts means shiny app for posts.
}
