% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{ssql}
\alias{ssql}
\title{Open Connections to the SteemSQL server. ssql means 'send to sql'.}
\usage{
ssql(uid = NA, pwd = NA)
}
\arguments{
\item{uid}{A character string of the SteemSQL ID.}

\item{pwd}{A character string of the SteemSQL password.}
}
\value{
A connection to the SeemSQL server.
}
\description{
Open Connections to the SteemSQL server. ssql means 'send to sql'.
}
\examples{
ssql()
}
