% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{steemsql_connection}
\alias{steemsql_connection}
\title{Open Connections to the SteemSQL server}
\usage{
steemsql_connection(sql_id = NA, sql_password = NA)
}
\arguments{
\item{sql_id}{A character string of the SteemSQL ID.}

\item{sql_password}{A character string of the SteemSQL password.}
}
\value{
A connection to the SeemSQL server
}
\description{
Open Connections to the SteemSQL server
}
\examples{
steemsql_connection()
}
