% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{tag_of_post}
\alias{tag_of_post}
\title{Retrieve tags from json_str of a post}
\usage{
tag_of_post(post_json_str)
}
\arguments{
\item{post_json_str}{json string with tag information}
}
\value{
tags
}
\description{
Retrieve tags from json_str of a post
}
