% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{unitconvert}
\alias{unitconvert}
\title{Concert unit from GV, MV, kV to numeric values.}
\usage{
unitconvert(x)
}
\arguments{
\item{x}{the unit GV, MV, kV to convert}
}
\value{
the numeric multiplier.
}
\description{
Concert unit from GV, MV, kV to numeric values.
}
