% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{vote}
\alias{vote}
\title{Vote report based on steemdb.com}
\usage{
vote(post = "https://steemit.com/cn/@dapeng/steemit-markdown")
}
\arguments{
\item{post}{complete link of a post on steemit}
}
\value{
a dataframe of a post's voter information
}
\description{
Vote report based on steemdb.com
}
\examples{
vote()
}
