% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{voter_sum}
\alias{voter_sum}
\title{Summary of the voters of a series of posts}
\usage{
voter_sum(mypost, if_plot = TRUE, top = 10)
}
\arguments{
\item{mypost}{A data frame of a series of posts retrieved with the appbase_api method}

\item{if_plot}{A logic value of whether plot the pie diagram.}

\item{top}{A numeric value of the Top voters plotted in the pie diagram.}
}
\value{
A data frame of the voters
}
\description{
Summary of the voters of a series of posts
}
