% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{who_not_vote}
\alias{who_not_vote}
\title{Find which followers have not voted a post yet}
\usage{
who_not_vote(postlink = "https://steemdb.com/cn/@dapeng/steemit-markdown")
}
\arguments{
\item{postlink}{character. the complete hyperlink to a target post.}
}
\value{
character. The name list of the followers who have not voted the target post yet.
}
\description{
Find which followers have not voted a post yet
}
\examples{
who_not_vote()
}
