% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{follower}
\alias{follower}
\title{Get a name list of an ID's followers and following}
\usage{
follower(id = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), 
    steemsql_connection)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{steemsql_connection}{A connection to the SteemSQL server.}
}
\value{
A name list of an ID's followers and following
}
\description{
Get a name list of an ID's followers and following
}
\examples{
follower("dapeng")
}
