% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{follower}
\alias{follower}
\title{Get a name list of an ID's followers and following}
\usage{
follower(id = "dapeng", method = c("steemdb.com", "steemdata.com"))
}
\arguments{
\item{id}{character without '@'}

\item{method}{the server to get data}
}
\value{
a name list of an ID's followers and following
}
\description{
Get a name list of an ID's followers and following
}
\examples{
follower()
}
