% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{following}
\alias{following}
\title{The id list of an id's following  from steemdb.com}
\usage{
following(id = "dapeng")
}
\arguments{
\item{id}{character without '@'}
}
\value{
character a dataframe of an ID's following info
}
\description{
The id list of an id's following  from steemdb.com
}
\examples{
following()
}
