% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gblog}
\alias{gblog}
\title{Rename the getBlog() function from the steemr2 package}
\usage{
gblog(id)
}
\arguments{
\item{id}{A Steem ID}
}
\value{
A dataframe with the ID's posts
}
\description{
Rename the getBlog() function from the steemr2 package
}
\examples{
gblog("dapeng")
}
