% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gur}
\alias{gur}
\title{Get the utopian review data from utopian.rocks api}
\usage{
gur(id = NA)
}
\arguments{
\item{id}{the steeian's ID to query}
}
\value{
a dataframe
}
\description{
Get the utopian review data from utopian.rocks api
}
\examples{
gur(NULL)
}
