% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{gvotep}
\alias{gvotep}
\title{A post's vote report based on steemdb.com. gvotep means 'get the vote information of a post.'}
\usage{
gvotep(postlink = NA)
}
\arguments{
\item{post}{A character string of the complete link of a post on steemit}
}
\value{
A dataframe of a post's voter information
}
\description{
A post's vote report based on steemdb.com. gvotep means 'get the vote information of a post.'
}
\examples{
gvotep("cn/@dapeng/steemit-markdown")
}
