% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{id_info}
\alias{id_info}
\title{Get an ID's detailed info}
\usage{
id_info(id = NA, method = c("steemdb.com", "steemsql.com", "appbase_api", 
    "steemdata.com"), steemsql_connection)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{steemsql_connection}{A connection to the SteemSQL server.}
}
\value{
A list or a dataframe of an ID's detailed info
}
\description{
Get an ID's detailed info
}
\examples{
id_info(id = "dapeng")
}
