% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{id_info}
\alias{id_info}
\title{Get an ID's detailed info}
\usage{
id_info(id = "dapeng", method = c("steemdb.com", "steemsql"))
}
\arguments{
\item{id}{character. Steem ID without '@'}

\item{method}{character. Steem data server.}
}
\value{
a dataframe of an ID's detailed info
}
\description{
Get an ID's detailed info
}
\examples{
id_info()
}
