% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{idlink}
\alias{idlink}
\title{Convert an id from a character to html hyperlink}
\usage{
idlink(id = "dapeng")
}
\arguments{
\item{id}{character without '@'}
}
\value{
a character
}
\description{
Convert an id from a character to html hyperlink
}
\examples{
idlink()
}
