% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{idlink}
\alias{idlink}
\title{Convert an id from a character to html hyperlink}
\usage{
idlink(id = "dapeng")
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}
}
\value{
A character string of the hyperlink to the Steem id.
}
\description{
Convert an id from a character to html hyperlink
}
\examples{
idlink()
}
