% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{post_df}
\alias{post_df}
\title{Get the detailed information of given posts from steemdb.com}
\usage{
post_df(postlinks)
}
\arguments{
\item{postlinks}{character. hyperlinks to target posts}
}
\value{
a dataframe of the detailed information of the given posts
}
\description{
Get the detailed information of given posts from steemdb.com
}
\examples{
post_df(c("cn/@dapeng/xuer-sale", "utopian-io/@dapeng/steemg-four-more"))
}
