% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{post_id}
\alias{post_id}
\title{Obtain an ID's post detailed info from steemdb.com}
\usage{
post_id(id = "dapeng", post_number = 3)
}
\arguments{
\item{id}{character without '@'}

\item{post_number}{numeric or NA. The number of the latest posts to be obtained. If NA, all the posts will be processed.}
}
\value{
data frame. an ID's post detailed info.
}
\description{
Obtain an ID's post detailed info from steemdb.com
}
\examples{
post_id()
}
