% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{post_id}
\alias{post_id}
\title{Obtain an ID's post detailed info from steemdb.com}
\usage{
post_id(id = NA, method = c("steemdb.com", "steemsql.com", "appbase_api", 
    "steemdata.com"), steemsql_connection, post_number = NA)
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{steemsql_connection}{A connection to the SteemSQL server.}

\item{post_number}{A numeric value or NA. The number of the latest posts to be obtained. If NA, all the posts will be processed.}
}
\value{
A data frame of an ID's post detailed info.
}
\description{
Obtain an ID's post detailed info from steemdb.com
}
\examples{
post_id("dapeng")
}
