% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{post_info}
\alias{post_info}
\title{Get the complete info of a single given post on steemdb.com}
\usage{
post_info(postlink = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), 
    steemsql_connection, selected = FALSE, newline = FALSE, oldcolname)
}
\arguments{
\item{postlink}{A character of the link to a post}

\item{method}{A character string of the Steem data server to connect.}

\item{steemsql_connection}{A connection to the SteemSQL server.}

\item{selected}{A logic value of whether return only selected info}

\item{newline}{A logic value of whether rbind the returned dataframe with an existing one}

\item{oldcolname}{A character string. If newline == TRUE, the returned dataframe is ordered accoring to oldcolname}
}
\value{
a dataframe
}
\description{
Get the complete info of a single given post on steemdb.com
}
\examples{
post_info("cn/@dapeng/steemit-markdown")
}
