% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{post_links}
\alias{post_links}
\title{Obtain an ID's post hyperlinks}
\usage{
post_links(id = NA, method = c("steemdb.com", "steemsql.com", "steemdata.com"), 
    steemsql_connection, post_number = 3, site = "steemit.com")
}
\arguments{
\item{id}{A character string of a Steem ID without '@'.}

\item{method}{A character string of the Steem data server to connect.}

\item{steemsql_connection}{A connection to the SteemSQL server.}

\item{post_number}{A numeric value or NA. The number of the latest posts to be obtained. If NA, the 100 latest posts will be processed.}

\item{site}{A character string of the site of the steem web UI}
}
\value{
A character string vector of an ID's post hyperlinks.
}
\description{
Obtain an ID's post hyperlinks
}
\examples{
post_links("dapeng")
}
