% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.R
\name{vote}
\alias{vote}
\title{A post's vote report based on steemdb.com}
\usage{
vote(post = "https://steemit.com/cn/@dapeng/steemit-markdown")
}
\arguments{
\item{post}{A character string of the complete link of a post on steemit}
}
\value{
A dataframe of a post's voter information
}
\description{
A post's vote report based on steemdb.com
}
\examples{
vote()
}
