% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steemr.r
\name{who_not_vote}
\alias{who_not_vote}
\title{Find which followers have not voted a post yet}
\usage{
who_not_vote(postlink = NA)
}
\arguments{
\item{postlink}{A character string of the link to a target post.}
}
\value{
A character vector of the name list of the followers who have not voted the target post yet.
}
\description{
Find which followers have not voted a post yet
}
\examples{
who_not_vote("cn/@dapeng/steemit-markdown")
}
